feat: allow defining iam role session name
This is a small enhancement that allows overriding the default
`assumed-from-leapp` role session name for AWS IAM Chained Sessions.

This is useful for organisations that share roles and alert on when
someone assumes them and use the role session name to differentiate
individuals.

This change adjusts the size of the main Electron window slightly to
accommodate the new field in the chained session setup page.