feat: allow configuring session duration
This feature allows a user to override the default environmental session
duration (currently 1200s).

Use case: Our organisation uses CloudWatch to alert when assuming
highly-privileged roles. The default session duration generates alerts
every 20 minutes during credential rotation. This feature allows the
user to set the session duration, and for us, reduce the number of
alerts generated.

Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
have the right to submit it under the open source license
indicated in the file; or

(b) The contribution is based upon previous work that, to the
best of my knowledge, is covered under an appropriate open
source license and I have the right under that license to
submit that work with modifications, whether created in whole
or in part by me, under the same open source license (unless
I am permitted to submit under a different license), as
Indicated in the file; or

(c) The contribution was provided directly to me by some other
person who certified (a), (b) or (c) and I have not modified
it.

(d) I understand and agree that this project and the contribution
are public and that a record of the contribution (including
all personal information I submit with it, including my
sign-off) is maintained indefinitely and may be redistributed
consistent with this project or the open source license(s)
involved.

Signed-off-by: Matthew Riedel <matthew.riedel@fluxfederation.com>