feat: make Web Console URL log out from current browser session (#520)
This change to the Web Console URL makes AWS log out from any existing
browser session, which avoids requiring a manual logout.

The implemented approach is based on [1] and its replies. It was found
during testing that the sign-in links require the `us-east-1` domain
prefix. Otherwise, the URL either redirects to the AWS Console landing
page, or returns a 400 error.

The session region set in Leapp is still considered, as it's part of the
`Destination` parameter.

I tested this change with AWS SSO and many AWS accounts, in multiple
regions for US and EU. I don't have access to any AWS GovCloud account,
so it would be interesting if anyone can test this change and validate
whether it works for GovCloud.

Fixes #309.

[1] https://serverfault.com/a/1097528

Signed-off-by: Michael Manganiello <adamantike@users.noreply.github.com>